more examples and details
I had imagined that the Mermaid class would pique people's interest making them check out the framework by reading the document, but seems I might have misjudged people's attention span. Despite being 4 days old and having 250+ views as of this writing, not a single fucker has responded, even with a "what the fuck is the use of this shot" on HN, so I decided to make things a bit more explicit.